+++ src/main/kotlin/icu/windea/pls/lang/editor/ParadoxInlineScriptEditorNotificationProvider.kt
@@ -26,7 +26,7 @@ class ParadoxInlineScriptEditorNotificationProvider : EditorNotificationProvider
         return Function f@{ fileEditor ->
             if (fileEditor !is TextEditor) return@f null
             val psiFile = file.toPsiFile(project) ?: return@f null
-            val message = PlsBundle.message("editor.notification.inlineScript", inlineScriptExpression)
+            val message = PlsBundle.message("editor.notification.inlineScript.text", inlineScriptExpression)
             val panel = EditorNotificationPanel(fileEditor, EditorNotificationPanel.Status.Info).text(message)
             panel.createActionLabel(PlsBundle.message("goto.usages")) {
                 GotoDeclarationAction.startFindUsages(fileEditor.editor, project, psiFile)